Consolidated log-sum-exp module (#95)
* Combine lse_implict and lse_scaled into a single module
* Consolidate test module too
