+++ gpfit/tests/t_logsumexp.py
@@ -1,8 +1,8 @@
-"unit tests for implicit log-sum-exp implementation"
+"unit tests for log-sum-exp functions"
 import unittest
 import numpy as np
 from numpy import array, arange
-from gpfit.lse_implicit import lse_implicit
+from gpfit.logsumexp import lse_implicit, lse_scaled
 
 SEED = 33404
 
@@ -32,6 +32,7 @@ class TestLSEimplicit1D(unittest.TestCase):
     def test_dydalpha_size(self):
         self.assertEqual(self.dydalpha.shape[0], self.x.shape[0])
 
+
 class TestLSEimplicit2D(unittest.TestCase):
     "tests with 2D input"
 
@@ -49,8 +50,44 @@ class TestLSEimplicit2D(unittest.TestCase):
     def test_dydalpha_shape(self):
         self.assertEqual(self.dydalpha.shape, self.x.shape)
 
-TESTS = [TestLSEimplicit1D,
-         TestLSEimplicit2D]
+
+class TestLSEScaled(unittest.TestCase):
+    "Test lse_implicit"
+
+    x = arange(1., 31.).reshape(15, 2)
+    alpha = 1.
+    y, dydx, dydalpha = lse_scaled(x, alpha)
+
+    def test_y_ndim(self):
+        self.assertEqual(self.y.ndim, 1)
+
+    def test_y_size(self):
+        self.assertEqual(self.y.size, self.x.shape[0])
+
+    def test_dydx_ndim(self):
+        self.assertEqual(self.dydx.ndim, 2)
+
+    def test_dydx_shape_0(self):
+        self.assertEqual(self.dydx.shape[0], self.x.shape[0])
+
+    def test_dydx_shape_1(self):
+        pass
+        # self.assertEqual(self.dydx.shape[0], ???????)
+
+    def test_dydalpha_ndim(self):
+        self.assertEqual(self.dydalpha.ndim, 1)
+
+    def test_dydalpha_size(self):
+        self.assertEqual(self.dydalpha.size, self.x.shape[0])
+
+    # test alpha is integer? negative? 0? array?
+
+
+TESTS = [
+    TestLSEimplicit1D,
+    TestLSEimplicit2D,
+    TestLSEScaled,
+]
 
 if __name__ == '__main__':
     SUITE = unittest.TestSuite()
